commands: move abandon code to abandon.rs
